ThreatFeed/requirements.txt
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
html5lib==1.1
idna==3.10
msal==1.33.0
Office365-REST-Python-Client==2.6.2
pycparser==2.23
PyJWT==2.10.1
pytz==2025.2
pywin32==311
requests==2.32.5
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
webencodings==0.5.1
